@@ -49,42 +52,51 @@
   "author": "Piotr Karwatka <pkarwatka@divante.pl>",
   "license": "MIT",
   "dependencies": {
-    "ajv": "^5.5.2",
+    "ajv": "^6.4.0",
+    "ajv-keywords": "^3.4.0",
     "body-parser": "^1.18.2",
-    "bodybuilder": "^2.2.10",
-    "command-router": "^1.0.1",
+    "bodybuilder": "2.2.13",
+    "commander": "^2.19.0",
     "compression": "^1.7.2",
     "config": "^1.30.0",
     "cors": "^2.8.4",
     "elasticdump": "^3.3.12",
-    "elasticsearch": "^13.3.1",
+    "elasticsearch": "^15.2.0",
+    "email-check": "^1.1.0",
     "express": "^4.16.3",
+    "graphql": "^0.10.1",
+    "graphql-tools": "^1.2.1",
     "humps": "^1.1.0",
-    "imageable": "^0.10.5",
     "jsonfile": "^4.0.0",
     "jwa": "^1.1.5",
     "jwt-simple": "^0.5.1",
     "kue": "^0.11.6",
     "lodash": "^4.17.10",
+    "mage2vuestorefront": "git+https://github.com/DivanteLtd/mage2vuestorefront.git",
+    "magento1-vsbridge-client": "github:DivanteLtd/magento1-vsbridge-client",
     "magento2-rest-client": "github:DivanteLtd/magento2-rest-client",
-    "migrate": "^1.5.0",
+    "merge-graphql-schemas": "^1.5.2",
+    "migrate": "^1.6.2",
     "mime-types": "^2.1.18",
     "morgan": "^1.9.0",
+    "nodemailer": "^4.6.8",
     "oauth-1.0a": "^1.0.1",
+    "pm2": "^2.10.4",
     "request": "^2.85.0",
+    "request-promise-native": "^1.0.5",
     "resource-router-middleware": "^0.6.0",
-    "soap": "^0.23.0",
-    "unirest": "^0.5.1",
-    "winston": "^2.4.2",
-    "pm2": "^2.10.4"    
+    "sharp": "^0.21.0",
+    "soap": "^0.25.0",
+    "winston": "^2.4.2"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-0": "^6.5.0",
-    "eslint": "^3.1.1",
-    "nodemon": "^1.17.4"
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.4.0",
+    "@babel/node": "^7.2.2",
+    "@babel/preset-env": "^7.4.2",
+    "apollo-server-express": "^1.3.6",
+    "eslint": "^4.16.0",
+    "nodemon": "^1.18.7"
   },
   "bugs": {
     "url": "https://github.com/DivanteLtd/vue-storefront-api/issues"
