@@ -1,29 +1,48 @@
 import resource from 'resource-router-middleware';
 import { apiStatus } from '../lib/util';
-const Ajv = require('ajv'); // json validator
+import { merge } from 'lodash';
+import PlatformFactory from '../platform/factory';
 
+const Ajv = require('ajv'); // json validator
+const fs = require('fs');
 const kue = require('kue');
 const jwa = require('jwa');
 const hmac = jwa('HS256');
+
+const _getProxy = (req, config) => {
+	const platform = config.platform
+	const factory = new PlatformFactory(config, req)
+	return factory.getAdapter(platform, 'order')
+};
+
 
 export default ({ config, db }) => resource({
 
 	/** Property name to store preloaded entity on `request`. */
 	id : 'order',
 
-	/** 
+	/**
 	 * POST create an order with JSON payload compliant with models/order.md
 	 */
 	create(req, res) {
 
 		const ajv = new Ajv();
-		const validate = ajv.compile(require('../models/order.schema.json'));
+		require('ajv-keywords')(ajv, 'regexp');
+
+		const orderSchema = require('../models/order.schema.js')
+		let orderSchemaExtension = {}
+		if(fs.existsSync('../models/order.schema.extension.json')) {
+			orderSchemaExtension = require('../models/order.schema.extension.json')
+		}
+		const validate = ajv.compile(merge(orderSchema, orderSchemaExtension));
 
 		if (!validate(req.body)) { // schema validation of upcoming order
 			console.dir(validate.errors);
-			apiStatus(res, validate.errors, 500);
+			apiStatus(res, validate.errors, 400);
 			return;
-		}				
+		}
+		const incomingOrder = { title: 'Incoming order received on ' + new Date() + ' / ' + req.ip, ip: req.ip, agent: req.headers['user-agent'], receivedAt: new Date(), order: req.body  }/* parsed using bodyParser.json middleware */
+		console.log(JSON.stringify(incomingOrder))
 
 		for (let product of req.body.products) {
 			let key = config.tax.calculateServerSide ? { priceInclTax: product.priceInclTax } : {  price: product.price }
@@ -33,7 +52,7 @@ export default ({ config, db }) => resource({
 				key.sku = product.sku
 			}
 			// console.log(key)
-			
+
 			if (!config.tax.usePlatformTotals) {
 				if (!hmac.verify(key, product.sgn, config.objHashSecret)) {
 					console.error('Invalid hash for ' + product.sku + ': ' + product.sgn)
@@ -43,10 +62,27 @@ export default ({ config, db }) => resource({
 			}
 		}
 
-		let queue = kue.createQueue(Object.assign(config.kue, { redis: config.redis }));
-		const job = queue.createJob('order', { title: 'Incoming order received on ' + new Date() + ' / ' + req.ip, ip: req.ip, agent: req.headers['user-agent'], receivedAt: new Date(), order: req.body  }/* parsed using bodyParser.json middleware */).save();
-		apiStatus(res, job.id, 200);
+		if (config.orders.useServerQueue) {
+			try {
+				let queue = kue.createQueue(Object.assign(config.kue, { redis: config.redis }));
+				const job = queue.create('order', incomingOrder).save( function(err){
+					if(err) {
+						console.error(err)
+						apiStatus(res, err, 500);
+					} else {
+						apiStatus(res, job.id, 200);
+					}
+				})
+			} catch (e) {
+				apiStatus(res, e, 500);
+			}
+		} else {
+			const orderProxy = _getProxy(req, config)
+			orderProxy.create(req.body).then((result) => {
+				apiStatus(res, result, 200);
+			}).catch(err => {
+				apiStatus(res, err.toString(), 500);
+			})
+		}
 	},
-
-	
 });
