@@ -21,7 +21,7 @@ cli.command('buildcache', () => {
           port: config.elasticsearch.port
         },
         log: 'debug',
-        apiVersion: '5.5',
+        apiVersion: config.elasticsearch.apiVersion,
         requestTimeout: 1000 * 60 * 60,
         keepAlive: false
       }
