@@ -1,32 +1,35 @@
 import { version } from '../../package.json';
 import { Router } from 'express';
 import order from './order';
-import img from './img';
-import catalog from './catalog'
-import user from './user'
-import stock from './stock'
-import cart from './cart'
-import product from './product'
-import sync from './sync'
+import catalog from './catalog';
+import user from './user';
+import stock from './stock';
+import review from './review';
+import cart from './cart';
+import product from './product';
+import sync from './sync';
 
 export default ({ config, db }) => {
 	let api = Router();
-	
+
 	// mount the catalog resource
 	api.use('/catalog', catalog({ config, db }) )
 
 	// mount the order resource
 	api.use('/order', order({ config, db }));
-	
+
 	// mount the user resource
 	api.use('/user', user({ config, db }));
 
 	// mount the stock resource
 	api.use('/stock', stock({ config, db }));
 
+	// mount the review resource
+	api.use('/review', review({ config, db }));
+
 	// mount the cart resource
 	api.use('/cart', cart({ config, db }));
-	
+
 	// mount the product resource
 	api.use('/product', product({ config, db }))
 
@@ -40,16 +43,22 @@ export default ({ config, db }) => {
 
 	/** Register the custom extensions */
 	for(let ext of config.registeredExtensions) {
+    let entryPoint
 
 		try {
-			let entryPoint = require('./extensions/' + ext)	
-			if (entryPoint) {
-				api.use('/ext/' + ext, entryPoint({ config, db }))
-				console.log('Extension ' + ext + ' registered under /ext/' + ext +' base URL')
-			}
+			entryPoint = require('./extensions/' + ext)
 		} catch (err) {
-			console.error(err)
-		}
+      try {
+        entryPoint = require(ext)
+      } catch (err) {
+        console.error(err)
+      }
+    }
+
+    if (entryPoint) {
+      api.use('/ext/' + ext, entryPoint({ config, db }))
+      console.log('Extension ' + ext + ' registered under /ext/' + ext +' base URL')
+    }
 	}
 
 	return api;
