@@ -1,28 +1,27 @@
 import { apiStatus, sgnSrc } from '../lib/util';
 import { Router } from 'express';
-import PlatformFactory from '../platform/factory'
+import PlatformFactory from '../platform/factory';
+
 const jwa = require('jwa');
 const hmac = jwa('HS256');
-
-const Ajv = require('ajv'); // json validator
 
 export default ({ config, db }) => {
 
 	let productApi = Router();
-	
+
 	const _getProxy = (req) => {
 		const platform = config.platform
 		const factory = new PlatformFactory(config, req)
 		return factory.getAdapter(platform,'product')
 	};
 
-	/** 
+	/**
 	 * GET get products info
 	 */
 	productApi.get('/list', (req, res) => {
 
 		const productProxy = _getProxy(req)
-		
+
 		if (!req.query.skus)
 			return apiStatus(res, 'skus parameter is required', 500);
 
@@ -33,17 +32,17 @@ export default ({ config, db }) => {
 		})
 	})
 
-	/** 
+	/**
 	 * GET get products info
 	 */
 	productApi.get('/render-list', (req, res) => {
 
 		const productProxy = _getProxy(req)
-		
+
 		if (!req.query.skus)
 			return apiStatus(res, 'skus parameter is required', 500);
 
-		productProxy.renderList(req.query.skus.split(','), req.query.currencyCode, req.query.storeId || 1).then((result) => {
+			productProxy.renderList(req.query.skus.split(','), req.query.currencyCode, (req.query.storeId && parseInt(req.query.storeId) > 0) ? req.query.storeId : 1).then((result) => {
 			result.items = result.items.map((item) => {
 				let sgnObj = item
 				if (config.tax.calculateServerSide === true) {
@@ -51,7 +50,7 @@ export default ({ config, db }) => {
 				} else {
 					sgnObj = {  price: item.price_info.extension_attributes.tax_adjustments.final_price }
 				}
-							
+
 				item.sgn = hmac.sign(sgnSrc(sgnObj, item), config.objHashSecret); // for products we sign off only price and id becase only such data is getting back with orders
 				return item
 			})
@@ -59,7 +58,7 @@ export default ({ config, db }) => {
 		}).catch(err=> {
 			apiStatus(res, err, 500);
 		})
-	})	
+	})
 
 	return productApi
 }
