@@ -15,34 +15,28 @@ class ProcessorFactory {
     this.config = app_config;
   }
 
-   getAdapter(entityType, indexName){
+   getAdapter(entityType, indexName, req, res){
 
     const moduleName = './' + entityType
 
-    if (!module_exists(moduleName))
-    {
+    if (!module_exists(moduleName)) {
       console.log('No additional data adapter for ' + entityType)
       return null
     }
 
     let adapter_class = require(moduleName);
-    if(!adapter_class)
-    {
+    if (!adapter_class) {
       console.log('No additional data adapter for ' + entityType)
       return null
-    }
-    else{
-      let adapter_instance = new adapter_class(this.config, entityType, indexName);
+    } else {
+      let adapter_instance = new adapter_class(this.config, entityType, indexName, req, res);
 
-      if((typeof adapter_instance.isValidFor == 'function') && !adapter_instance.isValidFor(entityType))
+      if ((typeof adapter_instance.isValidFor == 'function') && !adapter_instance.isValidFor(entityType))
         throw new Error('Not valid adapter class or adapter is not valid for ' + entityType);
 
       return adapter_instance;
-
     }
-
   }
-
 }
- 
+
 module.exports = ProcessorFactory;
